new e2e test for spread, and refactor affinity tests to share util methods
